Move actions to correct package
